feat(economy): add regret button to ladder
Users are able to revoke participation consent. However this will
blacklist them from joining the current season again. Revoke consent
has to be reset manually in the backend. People who regret joining
however are probably prone to not joining the next time.

Leave this as is. If someone explicitly complains about not being
able to rejoin we can reset on a per-case basis