feat(admissions): remove digital interview checbox
Removes the checkbox to select digital interview from the register
form in the ApplicantPortal. Currently there is not a desire
to support digital interviews as an option in this form. Applicants
will have to get in touch to arrange it themselves.

At the moment we just remove the UI component to toggle this
in case we want to revert this decision in the future